Make sure quasilinear mode also respects the upper bounds
This is achieved by disregarding values for s* and k* that would,
if were optimal, violate the proof of the upper bound.
In particular, if we decrease s* by an eps, the resulting solution
must be smaller than the upper bound.
Similarly, if we decrease k* by 1, the resulting solution
must be smaller than the upper bound.