Add WS and fix a crucial bug in the linear mode introduced in the previous commit
Weight separation, among other applications, allows implementing ramp secret sharing.
The bug in the linear mode used to lead to it always outputing a value close to the
upper bound and never actually doing the search.
This commit also introduces a primitive testing system that should help catch
regressions in the future.